新建应用

+++ nuwa/src/test/java/cn/asany/nuwa/template/service/ApplicationTemplateServiceTest.java
@@ -1,9 +1,12 @@
 package cn.asany.nuwa.template.service;
 
 import cn.asany.nuwa.TestApplication;
+import cn.asany.nuwa.app.bean.Routespace;
+import cn.asany.nuwa.app.service.RoutespaceService;
 import cn.asany.nuwa.template.bean.ApplicationTemplate;
 import lombok.extern.slf4j.Slf4j;
 import org.jfantasy.framework.jackson.JSON;
+import org.jfantasy.framework.util.common.ObjectUtil;
 import org.jfantasy.framework.util.common.file.FileUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,6 +26,8 @@ import java.util.List;
 class ApplicationTemplateServiceTest {
 
     @Autowired
+    private RoutespaceService routespaceService;
+    @Autowired
     private ApplicationTemplateService applicationTemplateService;
 
     @BeforeEach
@@ -39,8 +44,13 @@ class ApplicationTemplateServiceTest {
 
     @Test
     void deleteApplicationTemplate() throws IOException {
+        List<Routespace> routespaces = routespaceService.findAll();
         List<ApplicationTemplate> applicationTemplates = this.applicationTemplateService.findAll();
+
         for (ApplicationTemplate application : applicationTemplates) {
+            if (ObjectUtil.exists(routespaces, "applicationTemplate.id", application.getId())) {
+                continue;
+            }
             this.applicationTemplateService.deleteApplicationTemplate(application.getId());
         }
     }
